Avoid compiler warnings when printing off_t.
See:
http://stackoverflow.com/questions/586928/how-should-i-print-types-like-off-t-and-size-t

+++ stdcio.c
@@ -43,7 +43,8 @@ stdc_read(void* fd, const off_t offset, const size_t len, void* buf)
 static int
 stdc_read_dbg(void* fd, const off_t offset, const size_t len, void* buf)
 {
-  fprintf(stderr, "[ook] reading %zu bytes at %ld from %p\n", len, offset, fd);
+  fprintf(stderr, "[ook] reading %zu bytes at %jd from %p\n", len,
+          (intmax_t)offset, fd);
   return stdc_read(fd, offset, len, buf);
 }
 
@@ -63,7 +64,8 @@ stdc_write(void* fd, const off_t offset, const size_t len, const void* buf)
 static int
 stdc_write_dbg(void* fd, const off_t offset, const size_t len, const void* buf)
 {
-  fprintf(stderr, "[ook] writing %zu bytes at %ld into %p\n", len, offset, fd);
+  fprintf(stderr, "[ook] writing %zu bytes at %jd into %p\n", len,
+          (intmax_t)offset, fd);
   return stdc_write(fd, offset, len, buf);
 }
